requirements.txt
websockets>=11.0
cryptography>=41.0
python-dotenv>=1.0
uvloop>=0.17
